rename: `display` to `header` in wayland
