+++ src/platform_impl/linux/wayland/mod.rs
@@ -2,4 +2,4 @@
 // Copyright 2021-2023 Tauri Programme within The Commons Conservancy
 // SPDX-License-Identifier: Apache-2.0
 
-pub mod display;
+pub mod header;
